Build artist results and albums results components
