+++ src/Components/SearchResults.js
@@ -6,6 +6,8 @@ import { Collaborations } from '../Contexts/CollaborationsContextProvider'
 import '../Styles/SearchResults.scss'
 import { filterArtists, getNameFromURL } from '../utilities'
 import AddResults from './AddResults'
+import AlbumsResults from './AlbumsResults'
+import ArtistResults from './ArtistResults'
 
 const SearchResults = () => {
   const [ artists, setArtists ] = useState([])
@@ -26,10 +28,6 @@ const SearchResults = () => {
     }).catch(error => setErrorMessage(error))
   }, [location])
 
-  const handleClick = name => {
-    addCollaborator(name)
-  }
-
   const addCollaborator = name => {
     if (!collaborations.left.name) {
       setCollaborations({...collaborations, left: {name: name}})
@@ -39,68 +37,18 @@ const SearchResults = () => {
       setErrorMessage('There are already two musicians selected. Please remove one of the choices first.')
     }
   }
-  
-  // const addSearchResultsToCollaborators = () => {
-  //   addCollaborator(name.split('%20').join(' '))
-  // }
-
-  const artistCards = () => {
-    return artists.map(artist => {
-      const path = `/artist/${artist.id}`
-      return (
-        <div className="artist-name-container" key={artist.id}>
-          <Link to={path}>
-            <h3 className="artist-name">{artist.name}</h3>
-          </Link>
-          <button
-            className="add-collaborator-btn"
-            onClick={() => handleClick(artist.name)}
-          >choose</button>
-        </div>
-      )
-    })
-  }
-
-  const albumCards = () => {
-    return albums.map(album => {
-      return (
-        <div className="album-details" key={album.id}>
-          <div className="album-details-cover">
-            <img
-              src={album.cover}
-              alt={`Cover art for ${album.title}`}
-              className="album-cover"
-            />
-          </div>
-          <div className="album-details-container">
-            <Link to={`/album/${album.id}`}>
-              <h3 className="album-title">{album.title}</h3>
-            </Link>
-            <p className="album-artist">by {album.albumArtist}</p>
-          </div>
-        </div>
-      )
-    })
-  }
 
   return (
     <section className="search-results-page">
       <div className="search-reults-container">
         <section className="add-search-results">
-          <AddResults addCollaborator={addCollaborator}/>
+          <AddResults addCollaborator={addCollaborator} />
         </section>
         <section className="artist-results-container">
-          <AlbumsResults />
-          <h2 className="artist-results-header">Artists</h2>
-          <div className="artist-results">
-            { artistCards() }
-          </div>
+          <ArtistResults artists={artists} addCollaborator={addCollaborator} />
         </section>
         <section className="albums-results-container">
-          <h2 className="albums-results-header">Appears On: </h2>
-          <div className="albums-results">
-            { albumCards() }
-          </div>
+          <AlbumsResults albums={albums}/>
         </section>
       </div>
     </section>
